+++ src/commonMain/kotlin/number/AnyInt.kt
@@ -12,10 +12,8 @@ import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
-import kotools.types.ExperimentalSinceKotoolsTypes
 import kotools.types.Package
 import kotools.types.SinceKotoolsTypes
-import kotools.types.experimental.ExperimentalNumberApi
 import kotools.types.text.NotBlankString
 import kotools.types.text.toNotBlankString
 
@@ -42,11 +40,6 @@ public sealed interface AnyInt : Comparable<AnyInt> {
     /** Returns the string representation of this integer. */
     override fun toString(): String
 }
-
-/** Returns the negative of this integer. */
-@ExperimentalNumberApi
-@ExperimentalSinceKotoolsTypes("4.2")
-public operator fun AnyInt.unaryMinus(): Int = -toInt()
 
 /** Adds the [other] integer to this one. */
 @SinceKotoolsTypes("4.1")